% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneExp}
\alias{GeneExp}
\title{RNA-seq data matrix from GTEx}
\format{A matrix with 520 rows and 467 columns}
\source{
GETX URL:\href{https://www.gtexportal.org/home/}{
https://www.gtexportal.org/home/}.
}
\usage{
data("GeneExp")
}
\description{
GeneExp is a gene expression data matrix with rows (467 genes) and 520 columns. This data matrix is a subset of GTEx RNA-seq data which is a very big dataset. In this subset we only randomly chose 20 samples from each tissue type with more than 50 samples in the original GTEx data set.
}
\keyword{datasets}
