% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferTFact.R
\name{InferTFact}
\alias{InferTFact}
\title{Estimate TF activity score in a sample}
\usage{
InferTFact(exp, regnet)
}
\arguments{
\item{exp}{A numeric vector of gene expression levels for all TF target genes in a sample.}

\item{regnet}{A matrix, the network estimated by function \code{sciraInfNet}, with +1 referring to positive regulation, -1 negative regulation, and 0 no regulation.}
}
\value{
A vector storing the activity scores of all TFs in a specific sample.
}
\description{
\code{InferTFact} is an auxiliary function for function \code{sciraRegAct}.
}
\details{
\code{InferTFact} regresses the expression profiles of TF target genes against the binding profile of this TF on these genes. The output t-statistics are taken as the TF activity scores.
}
\examples{
# TF regulatory network with 100 genes and 5 TFs.
regnet.m <- matrix(sample(c(-1, 0, 1), 500, replace = TRUE), nrow = 100)
# gene expression vector (for one sample)
exp.v <- rnorm(100)
# TF activity score
TFact.v <- SCIRA:::InferTFact(exp.v, regnet.m)
}
