% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferTFact.R
\name{InferTFactPRL}
\alias{InferTFactPRL}
\title{Estimate TF activity score in different samples}
\usage{
InferTFactPRL(idx, tmp, regnet)
}
\arguments{
\item{idx}{A numeric vector indicating the samples from which you want to estimate the TF activity score.}

\item{tmp}{A numeric matrix of gene expression levels for all TF target genes, with rows referring to genes, columns to samples.}

\item{regnet}{The network estimated from function \code{sciraInfNet}, with +1 referring to positive regulation, -1 negative regulation, and 0 no regulation.}
}
\value{
A vector storing the activity scores of all TFs from user specified sample in the data matrix.
}
\description{
\code{InferTFactPRL } is an auxiliary function for function \code{sciraRegAct}.
}
\details{
The user needs to input the index of the sample in the data matrix from which \code{InferTFactPRL} estimates TF activity. \code{InferTFactPRL} accomplishes the work by regressing the expression profile of TF target genes against the binding profile of this TF on these genes. The output t-statistics are taken as the TF activity score.
}
\examples{
# TF regulatory network with 100 genes and 5 TFs.
regnet.m <- matrix(sample(c(-1, 0, 1), 500, replace = TRUE), nrow = 100)
# gene expression data matrix with 100 genes and 10 samples.
tmp.m <- matrix(rnorm(1000), nrow = 100)
# TF activity score
TFact.v <- scira:::InferTFactPRL(1, tmp.m, regnet.m)
}
