% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netLiver}
\alias{netLiver}
\title{Liver-specific regulatory network}
\format{A matrix, each column represent a TFs, and rows are target genes.}
\usage{
data("netLiver")
}
\description{
Liver-specific regulatory network inffered with sciraInfNet function, from GETx dataset.
}
\keyword{datasets}
