% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciraRegAct.R
\name{sciraRegAct}
\alias{sciraRegAct}
\title{Infer TF activity from single cell gene expression profile}
\usage{
sciraRegAct(data, regnet, norm = c("c", "z"), ncores = 4)
}
\arguments{
\item{data}{A gene expression data matrix, with rows referring to genes and columns to samples.}

\item{regnet}{A matrix, the regulatory network inferred from \code{sciraInfNet} function.}

\item{norm}{A character indicating the method used to normalize your input data set, "c" for "centering"; "z" for "z-score normalization".}

\item{ncores}{A numeric, the number of cores to use. See \code{mclapply}.}
}
\value{
A matrix of TF activity score with rows referring to TFs, columns to samples.
}
\description{
\code{sciraRegAct} calculates TF activity scores in user input data set. It could be a single cell gene expression dataset.
}
\details{
\code{sciraRegAct} is one of the two main functions in \code{SCIRA} package. It takes the output regulatory network from \code{sciraInfNet} as input, and computes the activity of all TFs in this network from user provided \code{data}.

The \code{data} matrix should be single cell gene expression data, with rows are genes and columns are samples. Duplicated row names are not allowed, so you should average the these rows before running \code{sciraRegAct}.

Note that it's very important that you use the same gene identifier through out the whole analysis.
}
\examples{
# estimate TF activity
# single cell RNA-seq dataset
data("scLiver")
# Lungnet
data("netLiver")
TFact <- sciraRegAct(scLiver,regnet=netLiver,norm="z",ncores=1)
}
