% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciraSelReg.R
\name{sciraSelReg}
\alias{sciraSelReg}
\title{Builds tissue-specific regulatory network}
\usage{
sciraSelReg(
  regnet.o,
  tissue.v,
  toi,
  cft.v = NULL,
  degth.v = rep(0.05, 1 + length(cft.v)),
  lfcth.v = c(1, rep(log2(1.5), length(cft.v)))
)
}
\arguments{
\item{regnet.o}{The output from 'sciraInfReg'}

\item{tissue.v}{A vector listing the tissue-type of each sample. Note, this vector must match the columns of the data matrix used as input to the 'sciraInfReg' function.}

\item{toi}{The tissue-type of interest.}

\item{cft.v}{A vector specifying the tissue types that should be treated as confounders. Often this will include blood since immune cells infiltrate most tissues}

\item{degth.v}{A vector specifying the adjusted P-value thresholds to use. The length of this vector equals 1+number of entries in 'cft.v'. The first P-value threshold refers to the comparison of the tissue type of interest to all other tissues together, whereas subsequent P-values are for the comparison of the tissue type of interest to each of the confounding tissue types.}

\item{lfcth.v}{A corresponding threshold on the log2 fold changes for each of the comparison. The length of this vector equals 1+number of entries in 'cft.v'.}
}
\value{
A list of 3 elements

netTOI
The tissue-specific regulon matrix with columns labeling the transcription factors and rows labeling the target genes, with entries either 1 (positive interaction), -1 (inhibitory interaction) or 0 (no association).

sumnet
A summary of the tissue-specific regulatory network listing number of targets and numbers of positive and inhibitory interactions

top
A list of matrices, each matrix listing the top-ranked DEGs for each comparison of interest. First entry corresponds to the comparison of the tissue type of interest against all other tissue-types, the remaining entries correspond to the comparison of the tissue type of interest to the confounding tissue types.
}
\description{
This function derives the tissue-specific regulatory network
}
\examples{
pheno.v <- c(rep(1,50),rep(2,50));
data.m <- matrix(rnorm(10000,0,1),nrow=1000,ncol=100);
data.m[1:100,1:50] <- matrix(rnorm(5000,2,1),nrow=100,ncol=50);
out.o <- LimmaFn(pheno.v,data.m);


}
