% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scLiver}
\alias{scLiver}
\title{scRNA-seq data matrix 
 
scRNA-seq dataset with 70 samples and 12875 genes, from Yang et al. 2017}
\format{A matrix with 12875 rows and 70 coloums.}
\usage{
data("scLiver")
}
\description{
scRNA-seq data matrix 
 
scRNA-seq dataset with 70 samples and 12875 genes, from Yang et al. 2017
}
\references{
Li Yang  Wei-Hua Wang  Wei-Lin Qiu  Zhen Guo  Erfei Bi  Cheng-Ran Xu. 
\emph{A single-cell transcriptomic analysis reveals precise pathways and regulatory mechanisms underlying hepatoblast differentiation}
Hepatology 11 (2017): 29353.
doi:\href{ https://doi.org/10.1002/hep.29353}{
10.1002/hep.29353}.
}
\keyword{datasets}
