% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFeid}
\alias{TFeid}
\title{Transcription factor genes annotated using Entrez gene ID}
\format{A vecto with 1385 observations. chr [1:1385] "9684" "59269" "2965" "6935" "9247" "3661" "6239" "11151" ...}
\usage{
data("TFeid")
}
\description{
TFeid is a vector of transcription factor Entrez IDs. This list of TFs is from MSigDB with 1385 TFs in it.
}
\keyword{datasets}
