\docType{data}
\name{tissue}
\alias{tissue}
\title{Tissue types of GTEX dataset}
\format{
A vector specifying the tissue-types of the 8555 GTEX bulk tissue samples
}
\usage{
data(tissue)
}
\description{
A vector specifying the tissue-types of the 8555 GTEX bulk tissue samples
}
\details{
Upon loading the data object, we find the following:
\itemize{
  \item \code{tissue.v}: the vector containing the tissue-types of the
  GTEX samples
}
}
\references{
Teschendorff AE, Wang, N.
\emph{Improved detection of tumor suppressor events in single-cell RNA-Seq data}
Submitted.
}
\keyword{datasets}
