% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimmaFn.R
\name{LimmaFn}
\alias{LimmaFn}
\title{Limma differential gene expression analysis for bulk tissue data}
\usage{
LimmaFn(pheno.v, data.m)
}
\arguments{
\item{pheno.v}{A phenotype vector. In our case, this will label the tissue-type in the bulk GTEX dataset}

\item{data.m}{The normalized gene expression data with rownames annotated to Entrez gene IDs (ensure these are unique) and columns labeling the bulk tissue samples.}
}
\value{
A list with two elements.

top
A list itself with as many entries as their contrasts, and each entry being a matrix of top-ranked DEGs.

cont
The matrix of contrasts.
}
\description{
This function runs limma analysis and outputs ranked list of DEGs for each contrast of interest
}
\examples{
pheno.v <- c(rep(1,50),rep(2,50));
data.m <- matrix(rnorm(10000,0,1),nrow=1000,ncol=100);
data.m[1:100,1:50] <- matrix(rnorm(5000,2,1),nrow=100,ncol=50);
out.o <- LimmaFn(pheno.v,data.m);

}
