\docType{data}
\name{tfEID}
\alias{tfEID}
\title{List of human transcription factors annotated to Entrez gene IDs}
\format{
A vector of Entrez gene IDs for 1385 human transcription factors
obtained from the Broad Institute's Molecular Signatures Database (MSigDB)
}
\usage{
data(TFeid)
}
\description{
A vector of Entrez gene IDs for 1385 human transcription factors
obtained from the Broad Institute's Molecular Signatures Database (MSigDB)
}
\details{
Upon loading the data object, we find the following:
\itemize{
  \item \code{TFeid.v}: the vector containing the Entrez gene IDs of the
  transcription factors
}
}
\references{
Teschendorff AE, Wang, N.
\emph{Improved detection of tumor suppressor events in single-cell
  RNA-Seq data}
Submitted.
}
\keyword{datasets}
